The program has been updated to have an ArrayList of type Sodas to hold the sodas, and it has a amount of money so that the customer can get change.
At line 100 we added an ArrayList for the different types of Sodas of type Sodas.

The userBuyingSoda() method was updated to have more options for the user when buying a soda, the user can now buy different sodas by selecting a menu number which corresponds to a soda, the sodas have an amount of stock, a price, and a name.

When the user buys a soda they will be allowed to choose the amount of sodas, then the total price will display. Followed by having the user enter a payment amount, the vending machine does not take more than $10. So if the user enters more than $10 they will be prompted to enter a valid amount.

The vending machine can run out of sodas and money, if the soda machine runs out of a selected item the user will be prompted "Sorry "soda name" is out of stock."

If the user enters an invalid quantity they will be prompted "Invalid quantity. Please enter a quantity between 1 and 3." depending on the soda and the amount available.